refactor(server): extract DJ queue rotation helpers into djHelpers.ts
Move advanceRotation, markTrackAsPlayed, removeDJFromQueue,
hasUnplayedTracks, findNextDJWithTracks out of DJQueueCommand.ts and
extract joinDjQueue from DJQueueJoinCommand.execute. Export
MAX_DJ_QUEUE_SIZE (re-exported from DJQueueCommand for compat) plus the
shared booth constants DJ_SLOT_SERVER_X / BEHIND_BOOTH_SERVER_Y.

Commands are now thin wrappers keeping their client.send(...) UI
notifications; helpers never depend on a Colyseus Client so the
upcoming NPC DJ manager can drive the same code paths. Zero behavior
change.
